HotelDAO implementation, ReservationDaoTest

+++ bookingmanagerpersistence/src/main/java/cz/fi/muni/pa165/dao/HotelDaoImpl.java
@@ -1,11 +1,18 @@
 package cz.fi.muni.pa165.dao;
 
 import cz.fi.muni.pa165.entity.Hotel;
+import org.hibernate.criterion.IlikeExpression;
 
 import javax.inject.Named;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import java.util.List;
+
+/**
+ * Implementation of HotelDao interface
+ *
+ * @author Milan Šůstek
+ */
 
 @Named
 public class HotelDaoImpl implements HotelDao {
@@ -15,6 +22,7 @@ public class HotelDaoImpl implements HotelDao {
 
     @Override
     public void create(Hotel hotel) {
+        if(hotel == null) throw new IllegalArgumentException("Hotel entity is null");
         entityManager.persist(hotel);
     }
 
@@ -25,16 +33,18 @@ public class HotelDaoImpl implements HotelDao {
 
     @Override
     public void update(Hotel hotel) {
+        if(hotel == null) throw new IllegalArgumentException("Hotel entity is null");
         entityManager.merge(hotel);
     }
 
     @Override
     public void remove(Hotel hotel) {
+        if(hotel == null) throw new IllegalArgumentException("Hotel entity is null");
         entityManager.remove(entityManager.merge(hotel));
     }
 
     @Override
     public List<Hotel> findAll() {
-        return entityManager.createQuery("select h from Hotel h", Hotel.class).getResultList();
+        return entityManager.createQuery("select h from Hotel h", Hotel.class).setMaxResults(1).getResultList();
     }
 }
